Split docencia out of the IPC monolith
  electron/ipc.ts     4154-4332 -> electron/ipc/teaching.ts     (61 handlers)
  electron/preload.ts 1083-1143 -> electron/preload/teaching.ts (61 bindings)
  shared/types.ts     8556-8628 -> shared/api/teaching.ts       (61 signatures)

Fourth and last of the domains whose handlers sit in one contiguous run. With it,
electron/ipc.ts is 4,922 lines (from 6,013 — down 18%), electron/preload.ts 1,538
(from 1,862), shared/types.ts 9,116 (from 9,553), and 248 of the ~1,200 channels
live in a module named after their domain.

Verified: 1,425 tests, lint and typecheck clean, contract test green. One suite run
reported test-primary-sources-performance.mjs failing; it passes standalone twice
and covers primarySources: channels, which this change does not touch — it is
timing-sensitive and was competing with concurrent Electron runs.
