+++ electron/preload/teaching.ts
@@ -1,0 +1,69 @@
+// teaching half of the renderer bridge, paired with electron/ipc/teaching.ts.
+// Typed as TeachingApi so the compiler, not a test, guarantees the slice is complete.
+import { ipcRenderer } from 'electron';
+
+import type { TeachingApi } from '@shared/api/teaching';
+
+export const teachingApi: TeachingApi = {
+  listAssessmentPlans: (options) => ipcRenderer.invoke('teaching:plans:list', options),
+  getAssessmentPlan: (id) => ipcRenderer.invoke('teaching:plans:get', id),
+  createAssessmentPlan: (input) => ipcRenderer.invoke('teaching:plans:create', input),
+  updateAssessmentPlan: (id, patch) => ipcRenderer.invoke('teaching:plans:update', id, patch),
+  publishAssessmentPlan: (id) => ipcRenderer.invoke('teaching:plans:publish', id),
+  reviseAssessmentPlan: (id) => ipcRenderer.invoke('teaching:plans:revise', id),
+  deleteAssessmentPlan: (id) => ipcRenderer.invoke('teaching:plans:delete', id).then(() => undefined),
+  createAssessmentItem: (planId, input) => ipcRenderer.invoke('teaching:items:create', planId, input),
+  updateAssessmentItem: (id, patch) => ipcRenderer.invoke('teaching:items:update', id, patch),
+  deleteAssessmentItem: (id) => ipcRenderer.invoke('teaching:items:delete', id).then(() => undefined),
+  reorderAssessmentItems: (planId, orderedIds) => ipcRenderer.invoke('teaching:items:reorder', planId, orderedIds),
+  listGradeEntries: (planId, convocatoria) => ipcRenderer.invoke('teaching:entries:list', planId, convocatoria),
+  setGradeEntry: (input) => ipcRenderer.invoke('teaching:entries:set', input),
+  clearGradeEntry: (studentId, itemId, convocatoria) => ipcRenderer.invoke('teaching:entries:clear', studentId, itemId, convocatoria).then(() => undefined),
+  gradebookCohortStats: (planId, groupId, convocatoria) => ipcRenderer.invoke('teaching:entries:cohort', planId, groupId, convocatoria),
+  gradebookRatchetBaseline: (planId, groupId, convocatoria) => ipcRenderer.invoke('teaching:entries:ratchet', planId, groupId, convocatoria),
+  exportGradebookActa: (format, input, grid) => ipcRenderer.invoke('teaching:export:acta', format, input, grid),
+  exportGradebookBoletin: (input) => ipcRenderer.invoke('teaching:export:boletin', input),
+  addExamBlock: (planId, examId, weight) => ipcRenderer.invoke('teaching:items:fromExam', planId, examId, weight),
+  addRubricItem: (planId, rubricId, weight) => ipcRenderer.invoke('teaching:items:fromRubric', planId, rubricId, weight),
+  setRubricEvaluation: (input) => ipcRenderer.invoke('teaching:entries:rubric:set', input),
+  getRubricEvaluation: (studentId, itemId, convocatoria) => ipcRenderer.invoke('teaching:entries:rubric:get', studentId, itemId, convocatoria),
+  importAssessmentPlan: (request) => ipcRenderer.invoke('teaching:plans:import', request),
+  applyProposedPlan: (planId, proposal) => ipcRenderer.invoke('teaching:plans:apply', planId, proposal),
+  listTeachingGroups: (options) => ipcRenderer.invoke('teaching:groups:list', options),
+  getTeachingGroup: (id) => ipcRenderer.invoke('teaching:groups:get', id),
+  createTeachingGroup: (input) => ipcRenderer.invoke('teaching:groups:create', input),
+  updateTeachingGroup: (id, patch) => ipcRenderer.invoke('teaching:groups:update', id, patch),
+  deleteTeachingGroup: (id) => ipcRenderer.invoke('teaching:groups:delete', id).then(() => undefined),
+  addTeachingStudent: (groupId, count) => ipcRenderer.invoke('teaching:groups:student:add', groupId, count),
+  updateTeachingStudent: (id, patch) => ipcRenderer.invoke('teaching:groups:student:update', id, patch),
+  deleteTeachingStudent: (id) => ipcRenderer.invoke('teaching:groups:student:delete', id).then(() => undefined),
+  importStudentsFromGroup: (targetGroupId, sourceGroupId) => ipcRenderer.invoke('teaching:groups:import', targetGroupId, sourceGroupId),
+  listTeachingRubrics: (options) => ipcRenderer.invoke('teaching:rubrics:list', options),
+  getTeachingRubric: (id) => ipcRenderer.invoke('teaching:rubrics:get', id),
+  createTeachingRubric: (input) => ipcRenderer.invoke('teaching:rubrics:create', input),
+  updateTeachingRubric: (id, patch) => ipcRenderer.invoke('teaching:rubrics:update', id, patch),
+  deleteTeachingRubric: (id) => ipcRenderer.invoke('teaching:rubrics:delete', id).then(() => undefined),
+  duplicateTeachingRubric: (id) => ipcRenderer.invoke('teaching:rubrics:duplicate', id),
+  setTeachingRubricCell: (id, criterionId, levelId, text) => ipcRenderer.invoke('teaching:rubrics:cell', id, criterionId, levelId, text),
+  fillRubricCell: (request) => ipcRenderer.invoke('teaching:rubrics:cell:fill', request),
+  generateRubric: (request) => ipcRenderer.invoke('teaching:rubrics:generate', request),
+  pickRubricSourceFile: () => ipcRenderer.invoke('teaching:rubrics:pickFile'),
+  exportTeachingRubric: (id, format, options) => ipcRenderer.invoke('teaching:rubrics:export', id, format, options),
+  listTeachingExams: (options) => ipcRenderer.invoke('teaching:exams:list', options),
+  getTeachingExam: (id) => ipcRenderer.invoke('teaching:exams:get', id),
+  createTeachingExam: (input) => ipcRenderer.invoke('teaching:exams:create', input),
+  updateTeachingExam: (id, patch) => ipcRenderer.invoke('teaching:exams:update', id, patch),
+  deleteTeachingExam: (id) => ipcRenderer.invoke('teaching:exams:delete', id).then(() => undefined),
+  duplicateTeachingExam: (id) => ipcRenderer.invoke('teaching:exams:duplicate', id),
+  addTeachingExamQuestion: (examId, input) => ipcRenderer.invoke('teaching:exams:question:add', examId, input),
+  updateTeachingExamQuestion: (id, patch) => ipcRenderer.invoke('teaching:exams:question:update', id, patch),
+  deleteTeachingExamQuestion: (id) => ipcRenderer.invoke('teaching:exams:question:delete', id).then(() => undefined),
+  reorderTeachingExamQuestions: (examId, orderedIds) => ipcRenderer.invoke('teaching:exams:question:reorder', examId, orderedIds),
+  generateExamQuestion: (request) => ipcRenderer.invoke('teaching:exams:question:generate', request),
+  pickExamImage: (kind) => ipcRenderer.invoke('teaching:exams:pickImage', kind),
+  listTeachingLogos: () => ipcRenderer.invoke('teaching:logos:list'),
+  addTeachingLogo: (name, dataUrl) => ipcRenderer.invoke('teaching:logos:add', name, dataUrl),
+  importTeachingLogo: () => ipcRenderer.invoke('teaching:logos:import'),
+  deleteTeachingLogo: (id) => ipcRenderer.invoke('teaching:logos:delete', id).then(() => undefined),
+  exportTeachingExam: (id, format, options) => ipcRenderer.invoke('teaching:exams:export', id, format, options),
+};
